+++ src/components/Home/about/About.jsx
@@ -5,7 +5,7 @@ import Image from "next/image";
 import { useTheme } from "../../../contexts/themeContext";
 import HeadingLinesAnimation from "../HeadingLinesAnimation/HeadingLinesAnimation";
 import { HomePkgsInBox } from "../../mui/HomePkgs";
-import { DecorativeBackgroundImage } from "../../Decorative/ItemBoxes";
+import { DecorativeBackgroundImage } from "../../Decorative/Decorative.style";
 
 export default function About() {
   const sectionRef = useRef(null);
